requirements.txt
Flask>=3.0
requests>=2.31
python-dotenv>=1.0
websocket-client>=1.6
